WIP: light refactor move oom handlers out of processwatcher
